requirements.txt
Flask==2.2.2
Cohere==4.0.0
Roboflow==0.2.20
opencv-python-headless==4.7.0.72
requests==2.31.0
google-generativeai==0.1.0
Gunicorn